fix(analytics): move the types inside the analytics types files
